Implement core scrollback functions
This is accomplished by replacing the `Vec` buffer with a `VecDeque`.

The *active region* represents the region in the buffer that is modifiable. It is essentially the latest `self.num_lines()` lines of the buffer.

Upon start up, the *visible region* is equal to the active region, and when new lines are added the visible/active region both move accordingly. However when you scroll, the visible region changes. When you type in new commands again tho, the visible region jumps back to equal the active region.