Fixed bug when scrolling by more than 1 line at a time.
It incorrectly reported `AtBottom` or `AtTop`, meaning that the dirty flag was not set and the terminal wasn't re-rendered.

Now the scroll logic is split into three cases: (a) its already at the top/bottom (b) after this movement it *will* be at the top/bottom (c) none of the above.